+++ test/artificial_samples/cwe_476.c
@@ -14,7 +14,7 @@ void func2(){
  free(data);
 }
 
-void main() {
+int main() {
 
   func1();
   func2();