+++ CHANGES.md
@@ -1,3 +1,8 @@
+dev
+====
+
+-   Added more documentation to checks
+
 0.2 (2019-06-25)
 =====