+++ CHANGES.md
@@ -7,6 +7,7 @@ dev
 -   Refactoring of logging and JSON support via --json (PR #30)
 -   Added file output support via --out (PR #30)
 -   Surpress logging of info, error and warning to STDOUT via --no-logging (PR #32)
+-   Added check-path feature via --check-path that searches paths between interesting input functions and cwe hits (PR #31)
 
 0.2 (2019-06-25)
 =====