+++ CHANGES.md
@@ -1,4 +1,4 @@
-dev
+0.3 (2019-12)
 ====
 
 -   Added more documentation to checks (PR #26)
@@ -8,8 +8,9 @@ dev
 -   Added file output support via --out (PR #30)
 -   Surpress logging of info, error and warning to STDOUT via --no-logging (PR #32)
 -   Added check-path feature via --check-path that searches paths between interesting input functions and cwe hits (PR #31)
+-   Added online documentation (PR #36, #37)
 -   Added convenience executable to enable shorter command line options (PR #40)
--   Added a plugin for integration into Ghidra (PR #42)
+-   Added a plugin for integration into Ghidra (PR #42, #43)
 
 0.2 (2019-06-25)
 =====